Fix for camera white balance
Former-commit-id: c57f716eaf91691218d916d619bb8ee3098a788b [formerly 174abf3f2f5b69befa615a7b19f66cf5ef2ca19b]
Former-commit-id: faa09d15cfd6eb1a04b48e798efd8456506b3f63 [formerly 53b0f7f27cf57ad43aa1aef5b9ac1d632af8eddd]
Former-commit-id: 7887ac64f32da726cd0f9c34f92b122285ec3dae
Former-commit-id: f0b01bb0e00837e9b1a13fc0e43bb420e7e1e67d

+++ Source/RawParser/Model/Effect/Balance.cs
@@ -107,7 +107,7 @@ namespace RawEditor.Effect
         }*/
 
 
-        public static void scaleColor(ref double r, ref double g, ref double b, double[] mul)
+        public static void scaleColor(ref double r, ref double g, ref double b, float[] mul)
         {
             r *= mul[0];
             g *= mul[1];
